+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 ## 0.85.0 (Unreleased)
+FEATURES:
+* storage: add `object_lock_configuration` attribute in `bucket` resource
+* storage: add `object_lock_legal_hold_status`, `object_lock_mode` and `object_lock_retain_until_date` attributes in `object` resource
 
 ## 0.84.0 (December 16, 2022)
 BUG FIXES: